pyproject.toml
[project]
name = "kbb"
version = "0.1.0"
description = "Autonomous Multi-Agent Knowledge Base Builder for RAG Systems"
authors = [{ name = "Alchemits' Squag 1", email = "squad1@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.12.2",
    "litellm==1.82.6",
    "fastmcp==3.2.0",
    "python-dotenv>=1.0.0",
    "langchain-text-splitters>=0.3.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=3.0.0",
    "einops>=0.8.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "html2text>=2025.4.15",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.10.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0"
]

[project.scripts]
kbb = "kbb.main:app"
run_crew = "kbb.main:app"
train = "kbb.main:train"
replay = "kbb.main:replay"
test = "kbb.main:test"
run_with_trigger = "kbb.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

[dependency-groups]
dev = [
    "coverage>=7.13.5",
    "mypy>=1.20.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.8",
]

